requirements.txt
affine>=2.4.0
geopandas>=0.13.2
numpy>=1.24.3
pandas>=2.0.2
Pillow>=9.5.0
pyproj>=3.6.0
rasterio>=1.3.7
setuptools>=58.1.0
Shapely>=2.0.1
tqdm>=4.65.0
